- Delete AdminController - Delete spring security tag
