+++ shoppingonline/.settings/oracle.eclipse.tools.webtier.ui.prefs
@@ -10,6 +10,7 @@
 //_hiddenCategory_\:http\://java.sun.com/jstl/sql_rt=false
 //_hiddenCategory_\:http\://java.sun.com/jstl/xml=false
 //_hiddenCategory_\:http\://java.sun.com/jstl/xml_rt=false
+//_hiddenCategory_\:http\://www.springframework.org/security/tags=false
 //_hiddenCategory_\:http\://www.springframework.org/tags=false
 //_hiddenCategory_\:http\://www.springframework.org/tags/form=false
 _hiddenCategory_\:HTML=false
