+++ shoppingonline/pom.xml
@@ -190,6 +190,11 @@
     	<artifactId>spring-security-ldap</artifactId>
     	<version>3.2.10.RELEASE</version>
 	</dependency>
+	<!-- dependency>
+    	<groupId>org.springframework.security</groupId>
+    	<artifactId>spring-security-taglibs</artifactId>
+    	<version>3.2.10.RELEASE</version>
+	</dependency -->
 	<dependency>
     	<groupId>org.apache.directory.server</groupId>
     	<artifactId>apacheds-core</artifactId>
