+++ shoppingonline/src/main/java/com/alithya/shoppingcart/controller/ShoppingOnLineAuthenticationController.java
@@ -26,7 +26,17 @@ public class ShoppingOnLineAuthenticationController {
 	public static final String SHOPPING_ONLINE_HOME_ADMIN = "shoppingonlinehomeadmin";
 	public static final String SHOPPING_ONLINE_CONNEXION = "shoppingonlineconnexion";
 	public static final String CONNEXION = "Access to your account";
-
+	public static final String REQUESTMAPPING_ADMIN = "/admin";
+	public static final String MODEL_NAME_ADMIN = "admin";
+	public static final String ADMINISTRATION ="You're Administrator !!!";
+		
+	@RequestMapping(value=REQUESTMAPPING_ADMIN, method=RequestMethod.GET)
+	public String getNewClient(ModelMap model){
+		
+		model.addAttribute(MODEL_NAME_ADMIN, ADMINISTRATION);
+		
+		return SHOPPING_ONLINE_HOME_ADMIN;
+	}
 	
 	@RequestMapping(value = REQUESTMAPPING_CONNECTION, method = RequestMethod.GET)
 	public String getConnection(ModelMap model, HttpServletRequest request) {
