+++ shoppingonline/src/main/webapp/WEB-INF/allJSPs/shoppingonlinebasket.jsp
@@ -645,6 +645,7 @@
     <div id="container">
         <header>
             <div id="headerCommandeMenu">
+                
                 <c:choose>
                 	<c:when test="${not empty connectionname }">
                 		<span id="saccueil"><a id="aaccueil" href="<spring:url value='/connect' />">Home</a></span>
@@ -652,7 +653,7 @@
                 	</c:when>
                 	<c:otherwise>
                 		<span id="saccueil"><a id="aaccueil" href="<spring:url value='/' />">Home</a></span>
-                		<span id="sconnexion" class="bouton"><a id="aconnexion" href='<spring:url value="/connection"></spring:url>'>Connection</a></span>
+                		<span id="sconnexion" class="bouton"><a id="aconnexion" href="<spring:url value='/makeconnection' />">Administration</a></span>
 					</c:otherwise>
 	              </c:choose>                
             </div>
